requirements.txt
line-bot-sdk
flask
matplotlib
gunicorn
imgurpython
pyimgur
numpy